% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emend-package.R
\docType{package}
\name{emend-package}
\alias{emend}
\alias{emend-package}
\title{emend: Cleaning Text Data with an AI Assistant}
\description{
Provides functions to clean and standardize messy data, including textual categories and free-text addresses, using Large Language Models. The package corrects typos, expands abbreviations, and maps inconsistent entries to standardized values. Ideal for Bioinformatics, business, and general data cleaning tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anuopensci/emend}
  \item \url{https://anuopensci.github.io/emend/}
  \item Report bugs at \url{https://github.com/anuopensci/emend/issues}
}

}
\author{
\strong{Maintainer}: Jiajia Li \email{lijia970324@gmail.com} (\href{https://orcid.org/0009-0003-7143-9336}{ORCID})

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
