% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{emend_what_language}
\alias{emend_what_language}
\title{Identify the language in the text.}
\usage{
emend_what_language(text, chat = get_default_chat())
}
\arguments{
\item{text}{A string or a factor that contains text information.}

\item{chat}{A chat object defined by ellmer}
}
\value{
A character vector of language names.
}
\description{
Identify the language in the text.
}
\examples{
\donttest{
chat <- ellmer::chat_ollama(model = "llama3.1:8b", seed = 0, echo = "none")
emend_what_language(c("\u733F\u3082\u6728\u304B\u3089\u843D\u3061\u308B", 
                      "\u4F60\u597D", "bon appetit"), chat = chat)
}

}
