% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_chat.R
\name{get_default_chat}
\alias{get_default_chat}
\title{Get or create the default chat object}
\usage{
get_default_chat()
}
\value{
A Chat object.
}
\description{
Get or create the default chat object
}
