% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{registration}
\alias{registration}
\title{Registration data for workshops}
\format{
A data frame with 221 rows and 3 columns:
\describe{
\item{Workshop}{Character: Workshop name and time.}
\item{Affiliation}{Character: Affiliation input by participants.}
\item{GroundTruth}{Character: The cleaned affiliation type, cleaned manually by author.}
}
}
\source{
Real registration data from the Biological Data Science Institute, Australian National University.
}
\usage{
registration
}
\description{
A dataset containing registration information for a series of workshops
The dataset includes columns for workshop name, affiliation, and the cleaned affiliation by hand.
}
\keyword{datasets}
