% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{emend_translate}
\alias{emend_translate}
\title{Translate text from one language to another.}
\usage{
emend_translate(text, to = "English", chat = get_default_chat())
}
\arguments{
\item{text}{The text to translate.}

\item{to}{The language to translate to. The default is "English".}

\item{chat}{An ellmer Chat object.}
}
\value{
A character vector of translated text.
}
\description{
Translate text from one language to another.
}
\examples{
chat <- ellmer::chat_ollama(model = "llama3.1:8b", seed = 0, echo = "none")
emend_translate(c("猿も木から落ちる", "你好", "bon appetit"), chat = chat)

}
